Update TFWriter to write files w/hclwrite package
In order to handle certian complex data structures and long or
multi-line strings, our current approach of converting to YAML flow
style strings before writing to HCL files would not longer suffice. This
commit implements writing with the
[hcl/hclwrite](https://pkg.go.dev/github.com/hashicorp/hcl/v2/hclwrite)
library which gives us more control over how HCL is written and enforces
style guidelines actively.

As part of this conversion, some of the data in the config needs to be
converted to cty.Value objects rather than interface{} for handling
dynamic typing. This all happens in the writer currently, however at
some point it may be better to rebuild the pipeline to work with
cty.Value from the start.