Embed resources in ghpc, update source format
The resources in the hpc-toolkit/resources directory are now embedded in
the binary and can be accessed by setting the source to resources/...,
whereas any local resource can be accessed by setting the source to
./..., ../... or /...

A temporary workaround was needed to get module info from tfconfig.
LoadModuleFromFS does not work with standard FS objects, and converting
is not possible. An issue has been filed with the project, but for now
the resource will be copied to a temp directory and read from there.