Move applyFunctions to Resources and rename
For simplicity, move applyFunctions into the YamlConfig so it can be
directly accessed during and after expand in the same way. This allows a
call to expand to produce a YAML that can be used again in create with
no differences in the final result.

The name was also changed to WrapSettingsWith to be more explanatory and
fit better under the Resource struct.