Fix bug in applied group terraform backends
A bug where the backends in all resource groups have the same prefix is
fixed with this change. Before the address of the default configuration
was being added to the groups and it's prefix was added. That prefix
would propogate as the defaults were added to other groups and therefore
the correct prefix was never updated.

In addition to the bug fix, the section of code was pulled out into it's
own function in expand as that better represents what it is doing.

Tests have been added to hopefully catch this type of bug in the future.