Default to not disabling services upon destroy
Alter default behavior of service enablement module to *not* disable
services upon destroy because it makes the implicit assumption that this
Terraform module is the sole user of services within the project. This
assumption is false in many circumstances and will cause terraform
destroy to fail when other resources are using an API (each VMs managed
outside of the module).