Change 'value' -> 'content' and update documentation
This commit makes two changes to the implementation of spack
environments.

The first is to change the keyword value to content, to help align
better with other modules of the toolkit (such as startup scripts).

The second is to enhance the description of the environments variable to
help users better understand how to use this functionality.