Avoid over-writing binary during integration
The integration tests infrequently fail with "text file busy" error
which arise when the ghpc binary is being invoked at the same time
a parallel copy of `ghpc build` is overwriting it. This commit
will run make only when binary does not exist or when an environment
variable, ALWAYS_RECOMPILE, is set to an value other than "no". The
default  value is "yes" which ensure that the behavior of the script
matches the current behavior unless the user opts in by setting the
variable to "no".