Extend periodic cleanup to reset Filestore API
Cloud Filestore has a limit on the number of internal networks that can
be created to peer the Filestore managed service to the user's VPCs. One
can avoid this problem by reusing names of Filestore instances
(undesirable if one wants to run builds in parallel) or periodically
resetting the Filestore API. This commit implements the reset.

https://cloud.google.com/filestore/docs/troubleshooting#system_limit_for_internal_resources_has_been_reached_error_when_creating_an_instance