Adding option to enable public ips
Through disable_login_public_ips in login node and disable_controller_public_ips in the controller.
It a weird semmantic, since it is true by default, but it was chosen
this way to maintain compatibility with V4.

By default, both the controller and the login node will have disable
public ips set to true. If set to false, and access_config is not
provided, it will generated a random public ip for each one of these

The partition doesn't seem to have the option available, and since it is
also typically less desirable, I did not add it.