Add install path variable to hybrid
A workaround to simplify hybrid deployment off controller. Rather than
have the full absolute path of the output_dir as the resume/suspend
program paths in cloud.conf, update the file with a sed command
replacing the output_dir with a provided install_dir, which acts
essentially as the prefix variable for the installation.

Alternatives attempted: Using a local_file block to read the data from
the generated cloud.conf file, use replace to update the content and
place the context back into the same file. This does not work as the
file does not exist at plan time, therefore terraform will fail.

Longer term solution: Build this into the setup_hybrid.py script used by
slurm-gcp in the underlying hybrid module.