Generalize install_dir update, trigger on vars
Generalizes the install_dir update to cloud.conf so that it can be
updated even on a second apply where the variable name changes. If
either output_dir or install_dir are changed, the replacement will be
triggered and it will replace whichever prefix existed before, it does
not have to perfectly match the output_dir variable.