Update ansible-lint pre-commit hook
The ansible-lint hook will run on every commit by default. This is due
to its not unreasonable assumption that the repo is structured primarily
for storing Ansible playbooks (roles, tasks, vars, etc.).

In practice, The Toolkit stores YAML files for blueprints and for
Ansible in separate directories and the Ansible is not structured the
way one would expect for a single playbook. This commit updates
pre-commit to trigger ansible linting only on changes to YAML files in
directories that do not contain blueprints.

Additionally, because ansible-lint will run on the entire repo when
triggered (regardless of the file that triggered it), we configure
ansible-lint itself to exclude paths that do not contain ansible.