batch login node supports multiple batch jobs
Update the variables to support passing in a list of batch jobs, or
using a list of batch jobs modules, with the Batch login node module.
All job templates will be placed in the same directory and instructions
for executing the job templates will be provided for all jobs. The first
provided job template will define the shape of the login node and it's
environment.