Add node_groups variable to partition module
This PR is the first in a set of changes that will fully implement node
groups in the slurm-gcp v5 partition module. To begin, the node_groups
variable has been added as a "in development"/preview option that will
behave exactly as the current module does if not used. If used, the
node_groups will be appended to the default partition currently
constructed in main.tf.