Adding gcsfuse to pre-existing-network-storage
This will allow mounting gcs buckets automatically through gcsfuse on
vm_instances. This should bring systems with automatic support to:
- Lustre
- NFS
- gcsfuse

Slurm V5 already supports the above systems and while V4 already
supports the same list, there is bug with gcsfuse in V4. It is scheduled
to be fixed in the next release.

Signed-off-by: Carlos Boneti <carlosboneti@google.com>