Modify Packer instructions to user
Packer does not have the equivalent of -chdir in Terraform. So we advise
the user to cd to the directory of the Packer deployment group. But we
do not advise them to return to the original working directory before
printing out Terraform instructions that assume one is in the original
working directory. Resolve by adding "cd -" to end of Packer output.