Inject build id into Omnia test, drop to last build
There have been known issues with the naming of the resources created by
the omnia example in automated testing. In order to allow parallel
builds, this PR injects the build ID into the name prefix before running
the test.

More thorough fixes are possible, but are not a priority for now. In
addition, the test has been dropped to last in the group to ensure that
all other tests can run if Omnia fails due to changes outside of our
teams control.