Changed DAOS community examples to use v0.3.0 terraform modules
https://daosio.atlassian.net/browse/DAOSGCP-145

The community/examples/intel/daos*.yaml blueprints use the Terraform
modules in the https://github.com/daos-stack/google-cloud-daos repo.

Changed the version tag of the modules referenced in the daos*.yaml
blueprints to v0.3.0 in order to use the latest terraform modules that
work with the DAOS v2.2 (the current release at this time).

Updated DAOS v2.0 references in the README.md to DAOS v2.2

Fixed incorrect replication factor setting in the README.md.
The Replication Factor used in the 'daos cont create' command was set
to 'rf:1'. That setting requires erasure coding which is not enabled
in the example deployment. As a result, the container will be created
successfully but an error will occur when attempting to create a file in
the container. Therefore, the 'rf:1' setting was changed to 'rf:0' to
prevent errors when following the instructions in the README.md

Signed-off-by: Mark Olson <115657904+mark-olson@users.noreply.github.com>