Reorder validator list to test blueprint correctness first
The test for a valid project ID had been set first, which would exit
before running other validators that depend on it. This PR simply moves
the one current validator that only check for correctness of the
blueprint ahead of the project ID validator to ensure it runs even if
the project ID is incorrect or credentials have not been setup yet.