Set minimal scopes for HTCondor IDTOKENs
The current solution creates tokens with full authorizations on each
node within the HTCondor pool. This PR continues to share the POOL
signing key with each node, but reduces the scope of each IDTOKEN that
is actually used in daemon-daemon authentication. It is anticipated that
future work will eliminate the dependency on a local copy of the POOL
signing key on each node.

- https://github.com/htcondor/htcondor/blob/8bcd2442756564ebbfdb6955fb16483806fda236/build/docker/services/README.md#using-the-execute-node-container
- https://github.com/htcondor/htcondor/blob/8bcd2442756564ebbfdb6955fb16483806fda236/build/docker/services/README.md#using-the-submit-access-point-container