Stop managing firewall on HTCondor pools
The goal of managing the firewall on HTCondor pools is to avoid regular
POSIX users escalating to use the service account needed by HTCondor
daemons. That is a reasonable objective, but is probably more confusing
for users than helpful. The solution is also not idempotent or
cross-platform friendly. We will return to this upon user demand.