Add health check for HTCondor VMs
Google Cloud Managed Instance Groups support a variety of health checks
that enable autohealing policies of VMs detected to be in an unhealthy
state. This commit adds a simple health check of port 9618 on any
machine within the Execute Point MIGs. TCP:9618 is the default "shared
port" configuration recommended for modern installations. It also adds
a single firewall rule to allow this check to run on any machine with an
HTCondor service account attached to it. This anticipates the
possibility of running Access Points or Central Managers in a Managed
Instance Group.