Enable OS Login by default in HTCondor execute points
Ensure that HTCondor execute points use the same default settings as the
vm-instance module to align user IDs in the job execution environment
with the uids in the job submission environment. Remove all vestiges of
(unused) metadata approach from examples and create a Known Issue in the
execute point README module.