Bugfix to HTCondor autoscaler script
When the autoscaler script is run on a machine other than the active
Access Point, the behavior of condor_q when run by the condor user
will be to print jobs only belonging to the condor user. This has the
unintended consequence of autoscaling the pool to 0 execute points.
Adding the -allusers flag ensures that all jobs are printed regardless
of where the script is run or the POSIX user it is run by (the condor
user is "special").