Fix broken test config use-resources.yaml
```
$ ./ghpc create tools/validate_configs/test_configs/use-resources.yaml -l ERROR --vars="project_id=io-playground"
module homefs uses module network1, but matching setting and outputs were not found. This may be because the value is set explicitly or set by a prior used module
module slurm_login uses module homefs, but matching setting and outputs were not found. This may be because the value is set explicitly or set by a prior used module
module slurm_login uses module scratchfs, but matching setting and outputs were not found. This may be because the value is set explicitly or set by a prior used module
module slurm_login uses module projectsfs, but matching setting and outputs were not found. This may be because the value is set explicitly or set by a prior used module
One or more used modules could not have their settings and outputs linked.
error: validator test_module_not_used failed
```