Resolve accidental destruction of startup-scripts
When the startup-script module is used with VM instance templates that
have create_before_destroy enabled (a common setting), a change to the
script contents will result in the script being created, then destroyed.
This commit resolves that by ensuring that all startup-script objects
are created with unique, repeatable identifiers. This allows the new
object to be created while destroying the old object with a different
identifier.