Resolve destroy/create race in vm-instance
If one changes a VM parameter that required destruction and replacement
of the VM (e.g. local_ssd_count) but does not change the zone of the VM,
then there is a race condition which causes the disk to be deleted along
with the VM (when var.auto_delete_boot_disk is true) but not re-created
because the explicit Terraform resource was not deleted. Thus the new VM
cannot be created until a 2nd Terraform apply refreshes state with the
now-deleted disk.

Resolve by using the initialize_params block instead of an explicit
google_compute_disk resource. Preserves behavior that setting
var.auto_delete_boot_disk=false allow a disk to be re-used when the VM
is destroyed and replaced.