Output multiple errors instead of first one
```yaml
  modules:
  - source: modules/network/vpc
    outputs: [gold, silver]

  - id: nw2
    source: modules/network/vpc
    settings:
      kill.switch: true
      kill@switch: true
      kill_switch: true
```

```shell
Error: a module id cannot be empty
34:   - source: modules/network/vpc
        ^

Error: requested output was not found in the module, module:  output: gold
35:     outputs: [gold, silver]
                  ^

Error: requested output was not found in the module, module:  output: silver
35:     outputs: [gold, silver]
                        ^

Error: a setting was added that is not found in the module
42:       kill_switch: true
                       ^

Error: a setting name contains a period, which is not supported; variable subfields cannot be set independently in a blueprint.
40:       kill.switch: true
                       ^

Error: a setting name must begin with a non-numeric character and all characters must be either letters, numbers, dashes ('-') or underscores ('_').
41:       kill@switch: true
                       ^
```