Support N>2 groups of HTCondor execute points
The solution currently only supports 2 groups of execute points because
the naming convention supports a fixed name with "spot" optionally
added. This commit modifies the approach to require user specification
of a group name that is unique to each group of execute points. There
cannot be a random component to the name otherwise a for_each will fail
in the htcondor-access-point module; future work may avoid need for user
to explicitly supply this value.