Guide user to solution when long family format
The upstream Slurm modules were written to need a special project ID
when referring to images by family. This project ID is not an actual
project ID and support for it was unintentionally dropped in prior work.
This commit adds validation to test when a user has provided this
special ID format and guides them to a solution.