Allow ramble setup module to use alternative forks for dependencies
This commit fixes an issue in the ramble-setup module where the
requirements.txt file was always downloaded from the same repo
(regardless of the repo used for the `git clone` call). As a result, if
a ref was passed that was not available in the main repo, this module
would fail.

After making this change, the ramble and spack template files were
unified more and added to the duplicate-diff tool.