Eliminate startup-script hasn't started message
The /etc/profile.d login prompt informational message makes the
assumption that the VM is running a startup-script that uses our
startup-script module. This assumption is broken when an image is built
using our startup-script module and then a VM is booted with that image
that does not execute our startup-script module. This assumption is also
broken upon reboots of Slurm VMs because our script is wrapped inside
a startup script solution developed by SchedMD that exits early when
Slurm has previously started successfully. We can reconsider enabling
this message more robustly as part of future work.