Improve MIG replacement policies for HTCondor Central Managers
Set the MIG replacement policy to PROACTIVE by default for Central
Managers. This ensures that configuration changes are propagated by
a terraform apply which updates the HTCondor configuration. This is safe
for Central Managers because they recover state dynamically through
periodic API calls to the rest of the cluster. Document the alternative
of OPPORTUNISTIC updates and how to manually trigger a MIG replacement.