HTCondor: expire ClassAds more rapidly
Decrease the default CLASSAD_LIFETIME from 15 minutes to 3 minutes. In
an on-premises system, allowing for long intervals between ClassAd
updates can be good to allow more machines to reboot. In the cloud, the
absence of ClassAd updates more likely indicates the intentional (or
automated) deletion of a VM. So it should be removed from the HTCondor
pool.