HTCondor autoscaler
Adopt a more conservative approach that the autoscaler should treat
nodes in any state that reflects automated MIG modification as an "idle"
node for the purposes of autoscaling. This helps prevent autoscaling
runaway when VMs are unable to enter the healthy state (which reflects
as "NONE" for currentAction in the MIG).