Add "submit" option to batch-job-template module
Adds an option "submit" (default false) that causes the deployment
to actually submit the job from the generated template, instead of
merely creating the template file.

Also changes the serverless-batch hello world example to use
submit: true and write a few files to the Filestore. The login node
from that example no longer generates a file to read, but the user
can log in to the login node and read the files generated by the
Batch job.