Support http_proxy in Windows GPU installation
Some customer environments need HTTP(S) proxies during execution of
Windows startup scripts. This commit adds that support by conditionally
templating setting a default proxy during execution of the script that
installs GPU drivers. Furthermore, it simplifies GPU installation to the
selection of a source URL for the script and argument to supply to
execute the script.