Update description of reservation_name for Slurm
The consumption of GCE reservations allows for 3 cases:
- SPECIFIC reservations, which must be named during the API call to
  create VMs
- automatically-consumed (ANY) reservations, which must not be named
  during the API call to create VMs
- refuse to use reservations at all

We presently support only the first 2 options by setting these values to
a reservation name (SPECIFIC) or to the empty string (ANY). If the user
specifies an ANY reservation, it will result in API failures. This
documentation update is intended to guide the user away from that
outcome. We anticipate future code changes to improve the experience for
case 2 and support case 3 at all.