Fix deployment of multiple Batch jobs
Three issues:

1. The submit_script_output_path did not depend on the
ID of the job to be submitted, so one job's submit script could
overwrite a different job's submit script.

2. Similarly, the job spec file name did not depend on the submit ID
of the job, and multiple jobs could overwrite each other if the job
ID or file name was not set explicitly in the blueprint.

3. The submit_job resource did not explicitly depend on the submit_script
resource and as a result would sometimes execute before the script
was written (causing submisison to fail with a file-not-found error).