Ensure local SSD solutions works upon reboot of Slurm nodes
When the local SSD mountpoint has not been mounted use SystemD to create
the RAID array and format it. This addresses the known behavior of the
Slurm-GCP solution in which it does not re-run startup-scripts upon
a power off/on (or reboot) cycle. During a typical power off/on cycle,
the local SSD contents are discarded and the disks must be re-assembled
and formatted.