Disallow local references to embedded modules
The use of source = "../.." notation within the Toolkit modules makes
the Toolkit a complete "package" of Terraform modules. This means the
whole set of Toolkit modules must be copied starting at the root
directory. However, any use of a leading ".." or "./" in the source line
in our YAML blueprint, will only stage the final directory into the
deployment folder.

This changes causes failure whenever a locally staged copy of a module
known to be embedded in the Toolkit is referenced. We anticipate
retiring embedding modules in favor of fully remote modules.