Add support for custom Docker daemon configuration
Docker daemons can be configured with a JSON file. Support for custom
data-roots is particularly valuable due to the performance impact of
storing Docker layers on performant storage. This commit adds support
custom JSON configuration along with the proper SystemD ordering for
data-root when it is specified. The assumption that data-root may be
provided by the local_ssd_filesystem option is made, but this dependency
is soft (After) and does not impose a strict requirement.