Drop validate flag to Docker daemon configuration
The current a3-highgpu-8g image does not support the --validate flag to
the dockerd daemon. In practice, this flag is really only useful in a
live system when a new configuration is overwriting an active
configuration (to prevent a bad update from propagating). In the Cloud,
most changes are being made at first boot so it always ends in failure
for the user when a bad configuration is provided.