Make gke-node-pool compatible with TPG 6.x
The gke-node-pool module uses older "attribute" syntax for the
GPU-related arguments that has been removed in the google Terraform
plugin 6.x. This commit replaces attribute syntax with block syntax.
The key to understanding this change is that a dynamic block iterating
over a list is equivalent to null when the list is empty (no dynamic
blocks are inserted).

The gpu_sharing_config and gpu_driver_installation_config settings are
not (and never were) list(object) in the Terraform plugin. They could only
ever taken on length 0 or 1. These are therefore being converted to object
format as they are in the API.

https://registry.terraform.io/providers/hashicorp/google/latest/docs/resources/container_cluster#nested_guest_accelerator
https://developer.hashicorp.com/terraform/language/attr-as-blocks