Add hybrid_conf variable
This variable contains all the needed values for a hybrid setup.
The values munge_secret and slurm_control_host are both compulsory.
Also for a correct hybrid cluster to work the slurm_cluster_name needs
to be provided to the module.
Adapted all the scripts and testing to this new variable.