Block broken release of nvidia-container-toolkit
The 1.17.7 release of nvidia-container-toolkit contains a regression
which breaks running GPU-enabled jobs under enroot in Slurm.

- https://github.com/NVIDIA/nvidia-container-toolkit/issues/1091
- https://github.com/NVIDIA/nvidia-container-toolkit/issues/1093
- https://github.com/NVIDIA/enroot/issues/232

While we wait for an updated package, this configuration will block
clusters from installing or upgrading to this package. If it is already
installed, this change does nothing. It should be forward-compatible in
the sense that it will not block new releases with sementically higher
versions.

This mitigates https://github.com/GoogleCloudPlatform/cluster-toolkit/issues/4144