Re-order runners in HCLS blueprint
Re-order runners in the HCLS blueprint so that the test sbatch scripts
are staged before Spack is built. Add features to the sbatch scripts
that tests for installation of Spack and Gromacs before continuing
execution. This improves the user experience by setting user
expectations for the potentially long build time of Gromacs.