Hold all nvidia software to the same version
Without this, during any combination of "update & upgrade", parts
of the nvidia software stack are liable to be upgraded and become
out of sync. While only libnvidia-compute-570-server causes immediate
errors, it is best to keep everything in sync with the image until
a point where an upgrade across all instances can be done.