feat(network): Add automatic deletion of hierarchical firewalls
When a VPC is destroyed, automatically created hierarchical firewall
policies can prevent the network from being deleted, causing the
`terraform destroy` operation to fail.

This commit introduces a new boolean variable,
`auto_delete_hierarchical_firewall_policies`, to the VPC module.
When this variable is set to `true`, a `null_resource` is created
that, upon destruction, will execute a script to delete any
firewall rules associated with the VPC.

This ensures that the VPC can be destroyed cleanly without manual
intervention to remove lingering firewall resources. This feature is
disabled by default.