Adding unkillable_step_timeout (UnkillableStepTimeut) as a default parameter to the H4D blueprint.
It has been observed that for large jobs on 32 node+ clusters, we often hit the QueuePair (QP) limit, leading to SIGKILL and subsequent cleaning of QPs.
The process of QP cleaning can take upto 12 minutes currently, which is more than twice the default UnkillableStepTimeout of 300 seconds.
This breach of timeout results in the affected nodes being marked as "DRAIN" rendering them unusable for any future jobs without explicit restart operations.

With this change we are providing an additional parameter within the blueprint to alter this value based on the workload and usage.
The change was tested on a 32 node cluster that ran Intel MPI benchmark and created >900K QPs before being killed. All the nodes successfully returned to "IDLE" state after QP cleaning

An earlier PR added SlurmdTimeout to the blueprint to mitigate this issue but intensive testing revealed that to be an incorrect approach, hence removing the slurmd_timeout(SlurmdTimeout)
parameter from the blueprint to avoid unnecessary clutter. Users can still specify slurmd_timeout in blueprints since it is still accepted as a valid blueprint param.